+++ library/oneview_firmware_bundle.py
@@ -31,7 +31,7 @@ notes:
 version_added: "2.3"
 requirements:
     - "python >= 2.7.9"
-    - "hpOneView >= 2.0.1"
+    - "hpeOneView >= 2.0.1"
 author: "Camila Balestrin (@balestrinc)"
 options:
     state: