+++ library/oneview_server_profile.py
@@ -31,7 +31,7 @@ description:
       automatically based on the server profile configuration if no server hardware was provided.
 version_added: "2.5"
 requirements:
-    - hpOneView >= 5.0.0
+    - hpeOneView >= 5.0.0
 author:
     - "Chakravarthy Racharla"
     - "Camila Balestrin (@balestrinc)"