+++ library/oneview_switch_facts.py
@@ -28,7 +28,7 @@ description:
 version_added: "2.3"
 requirements:
     - "python >= 2.7.9"
-    - "hpOneView >= 2.0.1"
+    - "hpeOneView >= 2.0.1"
 author: "Bruno Souza (@bsouza)"
 options:
     name: